requirements.txt
streamlit==0.87.0
tensorflow==2.6.0
Pillow==8.3.2
